feature-5: page promotionSearch and routes
