+++ src/pages/promotionForm/index.js
@@ -1,0 +1,15 @@
+import { useParams } from 'react-router-dom';
+
+
+const PromotionForm = () => {
+  const { id } = useParams();
+
+  return (
+    <div>
+      FORM
+      {id && <div>id: { id }</div>}
+    </div>
+  );
+};
+
+export default PromotionForm;
